move leq abstract tests to inversion test_abstract
